Merge branch 'be-meal-service-fix-meal-api' into 'be-meal-service

See merge request s08-final/S08P31S103!90

+++ back/healthone-meal-service/src/main/java/com/secui/healthone/domain/sportrecord/entity/SportRecord.java
@@ -23,7 +23,7 @@ public class SportRecord {
     @Column(name = "customsport_no")
     private Integer customSportNo;
     @Column(name = "sportrecord_worktime")
-    private Integer workTime;
+    private Float workTime;
     @Column(name = "sportrecord_consume_calorie")
     private Integer consumeCalorie;
     @Column(name = "sportrecord_heartrate")
